backend/requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0
fastapi[all]>=0.104.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0

